+++ animations/src/animations.c
@@ -10,7 +10,7 @@
  * 
  * delay_in_seconds: delay time
  */
-void delays(double delay_in_seconds)
+void delays(int delay_in_seconds)
 {
 	int millis = 1000 * delay_in_seconds;
 
@@ -24,9 +24,9 @@ void delays(double delay_in_seconds)
 void animate_world(int n, int world[][VIRTUAL_MAX_COLS + 2],
 				   int rows_count, int cols_count,
 				   int aux_world[][MAX_COLS],
-				   const int rule[RULE_SIZE], double delay_in_seconds)
+				   const int rule[RULE_SIZE], int delay_in_seconds)
 {
-	system("clear"); // CLear Screen in Unix
+	system("clear"); // Clear Screen in Unix
 
 	print_world(world, rows_count, cols_count);
